misunderstood test refactoring, https://github.com/technoweenie/faraday/pull/134#issuecomment-5087207
