+++ test/adapters/em_http_test.rb
@@ -6,5 +6,8 @@ module Adapters
     include Integration::Parallel
 
     def adapter; :em_http end
+
+    # https://github.com/eventmachine/eventmachine/pull/289
+    undef :test_timeout
   end
 end
