+++ test/adapters/em_synchrony_test.rb
@@ -6,5 +6,8 @@ module Adapters
     include Integration::Parallel
 
     def adapter; :em_synchrony end
+
+    # https://github.com/eventmachine/eventmachine/pull/289
+    undef :test_timeout
   end
 end
