+++ test/adapters/integration.rb
@@ -50,40 +50,6 @@ module Adapters
       end
     end
 
-    module Timeout
-      if Faraday::TestCase::LIVE_SERVER
-        def test_timeout
-          conn = create_connection(adapter, :request => {:timeout => 1, :open_timeout => 1})
-          assert_raise Faraday::Error::TimeoutError do
-            conn.get '/slow'
-          end
-        end
-      end
-    end
-
-    # https://github.com/dbalatero/typhoeus/issues/75
-    # https://github.com/toland/patron/issues/52
-    module GetWithBody
-      if Faraday::TestCase::LIVE_SERVER
-        def test_GET_with_body
-          response = create_connection(adapter).get('echo') do |req|
-            req.body = {'bodyrock' => true}
-          end
-          assert_equal %(get {"bodyrock"=>"true"}), response.body
-        end
-      end
-    end
-
-    # https://github.com/toland/patron/issues/9
-    # https://github.com/dbalatero/typhoeus/issues/84
-    module PutResponseHeaders
-      if Faraday::TestCase::LIVE_SERVER
-        def test_PUT_retrieves_the_response_headers
-          assert_match(/text\/html/, create_connection(adapter).put('echo_name').headers['content-type'])
-        end
-      end
-    end
-
     module Common
       def test_GET_retrieves_the_response_body
         assert_equal 'hello world', create_connection(adapter).get('hello_world').body
@@ -100,6 +66,18 @@ module Adapters
         response = create_connection(adapter).get('hello_world')
         assert_match(/text\/html/, response.headers['Content-Type'], 'original case fail')
         assert_match(/text\/html/, response.headers['content-type'], 'lowercase fail')
+      end
+
+      def test_GET_handles_headers_with_multiple_values
+        response = create_connection(adapter).get('multi')
+        assert_equal 'one, two', response.headers['set-cookie']
+      end
+
+      def test_GET_with_body
+        response = create_connection(adapter).get('echo') do |req|
+          req.body = {'bodyrock' => true}
+        end
+        assert_equal %(get {"bodyrock"=>"true"}), response.body
       end
 
       def test_POST_send_url_encoded_params
@@ -130,12 +108,29 @@ module Adapters
         assert_equal "file integration.rb text/x-ruby", resp.body
       end
 
+      def test_PUT_send_url_encoded_params
+        resp = create_connection(adapter).put do |req|
+          req.url 'echo_name'
+          req.body = {'name' => 'zack'}
+        end
+        assert_equal %("zack"), resp.body
+      end
+
       def test_PUT_send_url_encoded_nested_params
         resp = create_connection(adapter).put do |req|
           req.url 'echo_name'
           req.body = {'name' => {'first' => 'zack'}}
         end
         assert_equal %({"first"=>"zack"}), resp.body
+      end
+
+      def test_PUT_retrieves_the_response_headers
+        assert_match(/text\/html/, create_connection(adapter).put('echo_name').headers['content-type'])
+      end
+
+      def test_PATCH_send_url_encoded_params
+        resp = create_connection(adapter).patch('echo_name', 'name' => 'zack')
+        assert_equal %("zack"), resp.body
       end
 
       def test_OPTIONS
@@ -164,6 +159,13 @@ module Adapters
 
       def test_DELETE_retrieves_the_body
         assert_match(/deleted/, create_connection(adapter).delete('delete_with_json').body)
+      end
+
+      def test_timeout
+        conn = create_connection(adapter, :request => {:timeout => 1, :open_timeout => 1})
+        assert_raise Faraday::Error::TimeoutError do
+          conn.get '/slow'
+        end
       end
 
       def adapter
