+++ test/adapters/net_http_persistent_test.rb
@@ -4,7 +4,6 @@ module Adapters
   class NetHttpPersistentTest < Faraday::TestCase
     include Integration
     include Integration::NonParallel
-    include Integration::Timeout
 
     def adapter; :net_http_persistent end
   end
