+++ test/adapters/patron_test.rb
@@ -4,33 +4,18 @@ module Adapters
   class Patron < Faraday::TestCase
     include Integration
     include Integration::NonParallel
-    include Integration::GetWithBody
-    include Integration::PutResponseHeaders
-    include Integration::Timeout
 
     def adapter; :patron end
 
-    # https://github.com/toland/patron/issues/9
-    def test_PUT_send_url_encoded_params
-      resp = create_connection(adapter).put do |req|
-        req.url 'echo_name'
-        req.body = {'name' => 'zack'}
-      end
-      assert_equal %("zack"), resp.body
-    end
-
-    def test_PUT_send_url_encoded_nested_params
-      resp = create_connection(adapter).put do |req|
-        req.url 'echo_name'
-        req.body = {'name' => {'first' => 'zack'}}
-      end
-      assert_equal %({"first"=>"zack"}), resp.body
-    end
-
     # https://github.com/toland/patron/issues/34
-    def test_PATCH_send_url_encoded_params
-      resp = create_connection(adapter).patch('echo_name', 'name' => 'zack')
-      assert_equal %("zack"), resp.body
-    end
+    undef :test_PATCH_send_url_encoded_params
+
+    # https://github.com/toland/patron/issues/9
+    undef :test_PUT_retrieves_the_response_headers
+    undef :test_PUT_send_url_encoded_params
+    undef :test_PUT_send_url_encoded_nested_params
+
+    # https://github.com/toland/patron/issues/52
+    undef :test_GET_with_body
   end
 end
